+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## 2.0.1
+
+- Add `notifyEncryptionRequired` and `indicateEncryptionRequired` in CharacteristicProperties
+- `onBondStateChange` will return `BondState` enum instead of integer
+- Fix `isAdvertising` result on Android
+
 ## 2.0.0
 
 - Fix onReadRequest handler on android
 - BreakingChange: added deviceId in onReadRequest and onWriteRequest callbacks
-
 
 ## 1.0.0